# services/api-gateway/docs/swagger.yaml
swagger: "2.0"
info:
  description: API Gateway for Go Microservices Boilerplate
  title: Go Microservices API
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "1.0"
host: localhost:8080
basePath: /
schemes:
  - http
  - https

securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Enter your bearer token in the format: Bearer {token}"

paths:
  /api/v1/auth/login:
    post:
      summary: User login
      description: Login with email and password
      tags:
        - auth
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: credentials
          in: body
          required: true
          schema:
            $ref: "#/definitions/LoginRequest"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/LoginResponse"
        "401":
          description: Unauthorized
          schema:
            $ref: "#/definitions/ErrorResponse"

  /api/v1/users:
    get:
      summary: List users
      description: List users with pagination
      tags:
        - users
      security:
        - BearerAuth: []
      produces:
        - application/json
      parameters:
        - name: page
          in: query
          type: integer
          default: 1
          description: Page number
        - name: page_size
          in: query
          type: integer
          default: 20
          description: Page size
        - name: search
          in: query
          type: string
          description: Search query
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/ListUsersResponse"
    post:
      summary: Create a new user
      description: Create a new user account
      tags:
        - users
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: user
          in: body
          required: true
          schema:
            $ref: "#/definitions/CreateUserRequest"
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"

  /api/v1/users/{id}:
    get:
      summary: Get user by ID
      description: Get a user by their ID
      tags:
        - users
      security:
        - BearerAuth: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: User ID
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      summary: Update user
      description: Update user information
      tags:
        - users
      security:
        - BearerAuth: []
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: User ID
        - name: user
          in: body
          required: true
          schema:
            $ref: "#/definitions/UpdateUserRequest"
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/UserResponse"
        "400":
          description: Bad Request
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      summary: Delete user
      description: Delete a user by ID
      tags:
        - users
      security:
        - BearerAuth: []
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          type: string
          required: true
          description: User ID
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: "#/definitions/ErrorResponse"

  /auth/google:
    get:
      summary: Google OAuth2 login
      description: Redirect to Google OAuth2 login
      tags:
        - auth
      responses:
        "302":
          description: Redirect to Google

  /auth/github:
    get:
      summary: GitHub OAuth2 login
      description: Redirect to GitHub OAuth2 login
      tags:
        - auth
      responses:
        "302":
          description: Redirect to GitHub

definitions:
  CreateUserRequest:
    type: object
    required:
      - email
      - name
      - password
    properties:
      email:
        type: string
        example: user@example.com
      name:
        type: string
        example: John Doe
      password:
        type: string
        example: securePassword123

  UpdateUserRequest:
    type: object
    properties:
      name:
        type: string
        example: John Doe
      avatar_url:
        type: string
        example: https://example.com/avatar.png

  LoginRequest:
    type: object
    required:
      - email
      - password
    properties:
      email:
        type: string
        example: user@example.com
      password:
        type: string
        example: securePassword123

  LoginResponse:
    type: object
    properties:
      token:
        type: string
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      user:
        $ref: "#/definitions/UserResponse"

  UserResponse:
    type: object
    properties:
      id:
        type: string
        example: user-123
      email:
        type: string
        example: user@example.com
      name:
        type: string
        example: John Doe
      avatar_url:
        type: string
        example: https://example.com/avatar.png
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time

  ListUsersResponse:
    type: object
    properties:
      users:
        type: array
        items:
          $ref: "#/definitions/UserResponse"
      pagination:
        $ref: "#/definitions/PaginationResponse"

  PaginationResponse:
    type: object
    properties:
      page:
        type: integer
        example: 1
      page_size:
        type: integer
        example: 20
      total_count:
        type: integer
        example: 100
      total_pages:
        type: integer
        example: 5

  ErrorResponse:
    type: object
    properties:
      message:
        type: string
        example: Error message